asda

+++ src/js/03-feedback.js
@@ -3,26 +3,32 @@ import throttle from 'lodash.throttle';
 
 const form = document.querySelector('.feedback-form');
 const LOCALSTORAGE_KEY = 'feedback-form-state';
-const savedData = localStorage.getItem(LOCALSTORAGE_KEY);
-const parsedData = JSON.parse(savedData);
+const parsedData = JSON.parse(localStorage.getItem(LOCALSTORAGE_KEY));
 
 form.addEventListener('input', throttle(formData, 500));
+if (parsedData) {
+  form.elements.email.value = parsedData.email;
+  form.elements.message.value = parsedData.message;
+}
+
 function formData() {
   const data = {
     email: form.elements.email.value,
     message: form.elements.message.value,
   };
+
   localStorage.setItem(LOCALSTORAGE_KEY, JSON.stringify(data));
   //   console.log(parsedData.email);
 }
-form.elements.email.value = parsedData.email;
-form.elements.message.value = parsedData.message;
 
 form.addEventListener('submit', handleSubmit);
 
 function handleSubmit(evt) {
   evt.preventDefault();
-  console.log(parsedData);
+  if (form.elements.email.value === '' || form.elements.message.value === '') {
+    return alert('Please fill in all the fields!');
+  }
+  console.log(JSON.parse(localStorage.getItem(LOCALSTORAGE_KEY)));
   form.reset();
   localStorage.clear();
 }
